move cmd/api/abort to pkg/abort
As part of the sink refactor to use a gin http server, it would be helpful to share this code with the api server
