+++ cmd/api/auth/authn.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/kubearchive/kubearchive/cmd/api/abort"
+	"github.com/kubearchive/kubearchive/pkg/abort"
 	"github.com/kubearchive/kubearchive/pkg/cache"
 
 	"github.com/gin-gonic/gin"
