+++ cmd/api/auth/authz.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/kubearchive/kubearchive/cmd/api/abort"
+	"github.com/kubearchive/kubearchive/pkg/abort"
 	"github.com/kubearchive/kubearchive/pkg/cache"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
