+++ cmd/api/logging/logging.go
@@ -15,7 +15,7 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/kubearchive/kubearchive/cmd/api/abort"
+	"github.com/kubearchive/kubearchive/pkg/abort"
 	"github.com/kubearchive/kubearchive/pkg/files"
 	"github.com/ohler55/ojg/jp"
 	"github.com/ohler55/ojg/oj"
